# internal/model/user.go
package model

import "time"

type User struct {
	ID        int64     `db:"id" json:"id"`
	Login     string    `db:"login" json:"login"`
	Password  []byte    `db:"password" json:"password"`
	CreatedAt time.Time `db:"created_at" json:"created_at"`
	UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
}